SOO MANY CHANGES. Sorry, forgot to commit.

+++ src/com/sircular/circle/engine/Sprite.java
@@ -46,7 +46,7 @@ public class Sprite {
 		this.yvel += /*deltaAdjust(delta, */yaccel;//);
 	}
 	
-	public void update(int delta) {
+	public void update(long delta) {
 		
 	}
